+++ src/database/create-postgres-pool.ts
@@ -5,6 +5,9 @@ export function createPostgresPool(config: DBConfig) {
   return createPool(
     typeof config === 'string'
       ? config
-      : `postgres://${config.username}:${config.password}@${config.host}:${config.port}/${config.database}`
+      : `postgres://${config.username}:${config.password}@${config.host}:${config.port}/${config.database}`,
+    {
+      connectionTimeout: 'DISABLE_TIMEOUT',
+    }
   );
 }